use newer pure_python_packages format

+++ main/elegant_events/__dependencies__/__init__.py
@@ -50,7 +50,7 @@ def path_pieces(path):
     return [ *folders, filename, file_extension ]
 
 def remove(path):
-    if os.path.isdir(path):
+    if not Path(path).is_symlink() and os.path.isdir(path):
         shutil.rmtree(path)
     else:
         try:
@@ -58,12 +58,23 @@ def remove(path):
         except:
             pass
 
+def final_target_of(path):
+    # resolve symlinks
+    if os.path.islink(path):
+        have_seen = set()
+        while os.path.islink(path):
+            path = os.readlink(path)
+            if path in have_seen:
+                return None # circular broken link
+            have_seen.add(path)
+    return path
+    
 # 
 # globals
 # 
 this_file = make_absolute_path(__file__)
 this_folder = dirname(this_file)
-dependency_mapping_path = join(this_folder, '__dependency_mapping__.json')
+settings_path = join(this_folder, '..', 'settings.json')
 
 # 
 # find closest import path
@@ -99,13 +110,18 @@ if best_import_zone_match is None:
 import json
 from os.path import join
 # ensure it exists
-if not isfile(dependency_mapping_path):
-    with open(dependency_mapping_path, 'w') as the_file:
+if not isfile(settings_path):
+    with open(settings_path, 'w') as the_file:
         the_file.write(str("{}"))
-with open(dependency_mapping_path, 'r') as in_file:
-    dependency_mapping = json.load(in_file)
-    if not isinstance(dependency_mapping, dict):
-        raise Exception(f"""\n\n\nThis file is corrupt (it should be a JSON object):{dependency_mapping_path}""")
+with open(settings_path, 'r') as in_file:
+    settings = json.load(in_file)
+    if not isinstance(settings, dict):
+        raise Exception(f"""\n\n\nThis file is corrupt (it should be a JSON object):{settings_path}""")
+
+# ensure that pure_python_packages exists
+if not isinstance(settings.get("pure_python_packages", None), dict):
+    settings["pure_python_packages"] = {}
+dependency_mapping = settings["pure_python_packages"]
 
 # 
 # calculate paths
@@ -116,7 +132,7 @@ import_strings = []
 for dependency_name, dependency_info in dependency_mapping.items():
     counter += 1
     if dependency_name.startswith("__"):
-        raise Exception(f"""dependency names cannot start with "__", but this one does: {dependency_name}. This source of that name is in: {dependency_mapping_path}""")
+        raise Exception(f"""dependency names cannot start with "__", but this one does: {dependency_name}. This source of that name is in: {settings_path}""")
     
     target_path = join(this_folder, dependency_info["path"])
     relative_target_path = make_relative_path(to=target_path, coming_from=best_import_zone_match)
@@ -125,7 +141,8 @@ for dependency_name, dependency_info in dependency_mapping.items():
     eval_part = dependency_info.get("eval", dependency_name)
     unique_name = f"{dependency_name}_{random()}_{counter}".replace(".","")
     target_folder_for_import = join(this_folder, dependency_name)
-    if exists(target_folder_for_import):
+    if not Path(target_folder_for_import).is_symlink() or final_target_of(target_folder_for_import) != dependency_info["path"]:
+        # clear the way
         remove(target_folder_for_import)
-    # symlink the folder
-    Path(target_folder_for_import).symlink_to(dependency_info["path"])
+        # symlink the folder
+        Path(target_folder_for_import).symlink_to(dependency_info["path"])
